refactor: return a TypeConverter in TypeConverterFactory.createTypeConverter that can convert an object which has its class inherited from the source class
simplify redundant "? extends Object" declarations

move SourceTargetTypeConverterFactoryLookupTest to class under test's corresponding test package
